+++ test/babashka/pods/test_common.clj
@@ -34,6 +34,8 @@
                          3
                          true ;; local-date
                          true ;; roundtrip string array
+                         true ;; roundtrip metadata
+                         true ;; roundtrip metadata nested
                          1
                          "add the arguments"
                          nil